Fix Lv-010: 16-bit overflow in pre-warning timer assignment
Patches three sites in legacy man.c (lines 2744, 2780, 2880) where
`prewarningtimer * 60 * SECS_TO_MS_MULTIPLIER` was being computed in
16-bit unsigned arithmetic on the C2000, then widened to 32-bit only at
the assignment. With the production default prewarningtimer = 360, the
RHS evaluates to 4,320,000 which overflows mod 65,536 to 60,160; the
operator-configured 6-hour averaging window was actually firing after
~5 minutes.

Fix: cast the first operand to (unsigned long) so the multiplication
chain runs in 32-bit. With default settings:
  before: timers.prewarn = 60,160     (~5 min @ 5 ms/tick)
  after:  timers.prewarn = 4,320,000  (~6 h  @ 5 ms/tick)

This is a behavioural change. Operators see a longer, smoother
pre-warning baseline. Documented in:
  host/docs/CUSTOMER_NOTES.md  (operator-facing, "Behavioural changes
                                from bug fixes" section)
  host/docs/LEGACY_BUGS.md     (engineer-facing patch record)

Legacy .a00 still builds cleanly; binary grew from 445,962 to 445,975
bytes (+13 bytes for the wider arithmetic instructions).

+++ SourceCode1/SPU_Firmware/FirmwareSource_6.20/man.c
@@ -2741,7 +2741,11 @@ void PreWarningPowerUpInit(void)
 
 	ReadPrewarnStructfromNVMEM();
 	//reset the timer and carry on
-	timers.prewarn = allsensors.prewarningtimer * 60 * SECS_TO_MS_MULTIPLIER;//set in minutes
+	/* Lv-010 fix: cast forces 32-bit arithmetic so the multiplication
+	 * doesn't overflow the 16-bit `int` on TI C2000. With prewarningtimer=360
+	 * the RHS evaluates to 4,320,000 ticks (~6 h at 5 ms/tick); the previous
+	 * unmodified expression overflowed mod 65,536 to 60,160 (~5 min). */
+	timers.prewarn = (unsigned long)allsensors.prewarningtimer * 60UL * SECS_TO_MS_MULTIPLIER;//set in minutes
 }
 
 //Used to reset all prewarning conditions.
@@ -2773,7 +2777,11 @@ void PreWarningInit(void)
 
 	prewarn.prewarncompleteflag = TRUE;	//pre warning 6 hour averaging completed when this flag is set to true
 	PrewarnReset();
-	timers.prewarn = allsensors.prewarningtimer * 60 * SECS_TO_MS_MULTIPLIER;//set in minutes
+	/* Lv-010 fix: cast forces 32-bit arithmetic so the multiplication
+	 * doesn't overflow the 16-bit `int` on TI C2000. With prewarningtimer=360
+	 * the RHS evaluates to 4,320,000 ticks (~6 h at 5 ms/tick); the previous
+	 * unmodified expression overflowed mod 65,536 to 60,160 (~5 min). */
+	timers.prewarn = (unsigned long)allsensors.prewarningtimer * 60UL * SECS_TO_MS_MULTIPLIER;//set in minutes
 	// PF Gen 2 - 07-Mar-12 Old alarm system backend removal - memset(eventstring,0,EVENTMSG_LEN_INBYTES_PART2);
 	// PF Gen 2 - 07-Mar-12 Old alarm system backend removal - LogEvent(PREWARN_INITIATED,HIGH_PRIORITY,NO,eventstring);	//no extra message to pass in for this event
 	SetBitEvent(PREWARN_INIT_EVENT); // PF Gen 2 - 07-Mar-12 Old alarm system backend removal - prewarn reset
@@ -2869,7 +2877,11 @@ int prewarnval;
 				prewarn.calcacc[i] = 0;	
 				prewarn.points[i] = 0;
 			}
-			timers.prewarn = allsensors.prewarningtimer * 60 * SECS_TO_MS_MULTIPLIER;//set in minutes
+			/* Lv-010 fix: cast forces 32-bit arithmetic so the multiplication
+	 * doesn't overflow the 16-bit `int` on TI C2000. With prewarningtimer=360
+	 * the RHS evaluates to 4,320,000 ticks (~6 h at 5 ms/tick); the previous
+	 * unmodified expression overflowed mod 65,536 to 60,160 (~5 min). */
+	timers.prewarn = (unsigned long)allsensors.prewarningtimer * 60UL * SECS_TO_MS_MULTIPLIER;//set in minutes
 			WritePrewarnStructtoNVMEM();
 		}
 	}
